@@ -4,12 +4,12 @@ plugins {
 
 android {
     namespace 'com.r00205604.calendar'
-    compileSdk 32
+    compileSdk 33
 
     defaultConfig {
         applicationId "com.r00205604.calendar"
         minSdk 29
-        targetSdk 32
+        targetSdk 33
         versionCode 1
         versionName "1.0"
 
