@@ -1,5 +1,6 @@
 package com.r00205604.calendar;
 
+import android.graphics.Color;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -7,40 +8,59 @@ import android.view.ViewGroup;
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
 
+import java.time.LocalDate;
 import java.util.ArrayList;
 
-public class CalendarAdapter extends RecyclerView.Adapter<CalendarViewHolder> {
-
-    private final ArrayList<String> daysOfMonth;
+class CalendarAdapter extends RecyclerView.Adapter<CalendarViewHolder>
+{
+    private final ArrayList<LocalDate> days;
     private final OnItemListener onItemListener;
 
-    public CalendarAdapter(ArrayList<String> daysOfMonth, OnItemListener onItemListener) {
-        this.daysOfMonth = daysOfMonth;
+    public CalendarAdapter(ArrayList<LocalDate> days, OnItemListener onItemListener)
+    {
+        this.days = days;
         this.onItemListener = onItemListener;
     }
 
     @NonNull
     @Override
-    public CalendarViewHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
+    public CalendarViewHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType)
+    {
         LayoutInflater inflater = LayoutInflater.from(parent.getContext());
         View view = inflater.inflate(R.layout.calendar_cell, parent, false);
         ViewGroup.LayoutParams layoutParams = view.getLayoutParams();
-        layoutParams.height = (int) (parent.getHeight() * 0.166666666);
-        return new CalendarViewHolder(view, onItemListener);
+        if(days.size() > 15) //month view
+            layoutParams.height = (int) (parent.getHeight() * 0.166666666);
+        else // week view
+            layoutParams.height = (int) parent.getHeight();
+
+        return new CalendarViewHolder(view, onItemListener, days);
     }
 
     @Override
-    public void onBindViewHolder(@NonNull CalendarViewHolder holder, int position) {
-        holder.dayOfMonth.setText(daysOfMonth.get(position));
+    public void onBindViewHolder(@NonNull CalendarViewHolder holder, int position)
+    {
+        final LocalDate date = days.get(position);
+
+        holder.dayOfMonth.setText(String.valueOf(date.getDayOfMonth()));
+
+        if(date.equals(CalendarUtils.selectedDate))
+            holder.parentView.setBackgroundColor(Color.LTGRAY);
+
+        if(date.getMonth().equals(CalendarUtils.selectedDate.getMonth()))
+            holder.dayOfMonth.setTextColor(Color.BLACK);
+        else
+            holder.dayOfMonth.setTextColor(Color.LTGRAY);
     }
 
     @Override
-    public int getItemCount() {
-        return daysOfMonth.size();
+    public int getItemCount()
+    {
+        return days.size();
     }
 
     public interface  OnItemListener
     {
-        void onItemClick(int position, String dayText);
+        void onItemClick(int position, LocalDate date);
     }
 }
