@@ -6,20 +6,28 @@ import android.widget.TextView;
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
 
-public class CalendarViewHolder extends RecyclerView.ViewHolder implements View.OnClickListener {
+import java.time.LocalDate;
+import java.util.ArrayList;
 
+public class CalendarViewHolder extends RecyclerView.ViewHolder implements View.OnClickListener
+{
+    private final ArrayList<LocalDate> days;
+    public final View parentView;
     public final TextView dayOfMonth;
     private final CalendarAdapter.OnItemListener onItemListener;
-
-    public CalendarViewHolder(@NonNull View itemView, CalendarAdapter.OnItemListener onItemListener) {
+    public CalendarViewHolder(@NonNull View itemView, CalendarAdapter.OnItemListener onItemListener, ArrayList<LocalDate> days)
+    {
         super(itemView);
+        parentView = itemView.findViewById(R.id.parentView);
         dayOfMonth = itemView.findViewById(R.id.cellDayText);
         this.onItemListener = onItemListener;
         itemView.setOnClickListener(this);
+        this.days = days;
     }
 
     @Override
-    public void onClick(View view) {
-        onItemListener.onItemClick(getAdapterPosition(), (String) dayOfMonth.getText());
+    public void onClick(View view)
+    {
+        onItemListener.onItemClick(getAdapterPosition(), days.get(getAdapterPosition()));
     }
 }
