@@ -1,84 +1,82 @@
 package com.r00205604.calendar;
 
+import static com.r00205604.calendar.CalendarUtils.daysInMonthArray;
+import static com.r00205604.calendar.CalendarUtils.monthYearFromDate;
+
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.recyclerview.widget.GridLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
+import android.content.Intent;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.TextView;
 
 import java.time.LocalDate;
-import java.time.YearMonth;
-import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
 
-public class MainActivity extends AppCompatActivity implements CalendarAdapter.OnItemListener {
+public class MainActivity extends AppCompatActivity implements CalendarAdapter.OnItemListener
+{
+    private TextView monthYearText;
+    private RecyclerView calendarRecyclerView;
 
-    private TextView monthYearTV;
-    private RecyclerView calendarRV;
-    private LocalDate selectedDate;
 
     @Override
-    protected void onCreate(Bundle savedInstanceState) {
+    protected void onCreate(Bundle savedInstanceState)
+    {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
 
-        calendarRV = findViewById(R.id.calendarRV);
-        monthYearTV = findViewById(R.id.monthYearTV);
-        selectedDate = LocalDate.now();
+        calendarRecyclerView = findViewById(R.id.calendarRV);
+        monthYearText = findViewById(R.id.monthYearTV);
+
+        CalendarUtils.selectedDate = LocalDate.now();
         setMonthView();
     }
 
-
-
-    private void setMonthView() {
-        monthYearTV.setText(monthYearFromDate(selectedDate));
-        ArrayList<String> daysInMonth = daysInMonthArray(selectedDate);
+    private void setMonthView()
+    {
+        monthYearText.setText(monthYearFromDate(CalendarUtils.selectedDate));
+        ArrayList<LocalDate> daysInMonth = daysInMonthArray();
 
         CalendarAdapter calendarAdapter = new CalendarAdapter(daysInMonth, this);
         RecyclerView.LayoutManager layoutManager = new GridLayoutManager(getApplicationContext(), 7);
-
-        calendarRV.setLayoutManager(layoutManager);
-        calendarRV.setAdapter(calendarAdapter);
+        calendarRecyclerView.setLayoutManager(layoutManager);
+        calendarRecyclerView.setAdapter(calendarAdapter);
     }
 
-    private ArrayList<String> daysInMonthArray(LocalDate date) {
-        ArrayList<String> daysInMonthArray = new ArrayList<>();
-        YearMonth yearMonth = YearMonth.from(date);
-
-        int daysInMonth = yearMonth.lengthOfMonth();
-
-        LocalDate firstOfMonth = selectedDate.withDayOfMonth(1);
-        int dayOfWeek = firstOfMonth.getDayOfWeek().getValue();
-
-        for(int i = 1; i <= 42; i++) {
-            if(i <= dayOfWeek || i > daysInMonth + dayOfWeek) {
-                daysInMonthArray.add("");
-            } else {
-                daysInMonthArray.add(String.valueOf(i - dayOfWeek));
-            }
-        }
-        return  daysInMonthArray;
-    }
-
-    private String monthYearFromDate(LocalDate date) {
-        DateTimeFormatter formatter = DateTimeFormatter.ofPattern("MMMM yyyy");
-        return date.format(formatter);
-    }
-
-    public void lastMonth(View view) {
-        selectedDate = selectedDate.minusMonths(1);
+    public void lastMonth(View view)
+    {
+        CalendarUtils.selectedDate = CalendarUtils.selectedDate.minusMonths(1);
         setMonthView();
     }
 
-    public void nextMonth(View view) {
-        selectedDate = selectedDate.plusMonths(1);
+    public void nextMonth(View view)
+    {
+        CalendarUtils.selectedDate = CalendarUtils.selectedDate.plusMonths(1);
         setMonthView();
     }
 
     @Override
-    public void onItemClick(int position, String dayText) {
+    public void onItemClick(int position, LocalDate date)
+    {
+        if(date != null)
+        {
+            CalendarUtils.selectedDate = date;
+            setMonthView();
+        }
+    }
 
+    public void weeklyAction(View view)
+    {
+        startActivity(new Intent(this, WeekViewActivity.class));
     }
 }
+
+
+
+
+
+
+
+
